requirements.txt
tensorflow==2.19.0
keras==3.10.0
flask
gunicorn
numpy
pillow
